# docs/openapi/c2c-market-api-v1.yaml
openapi: 3.1.0
info:
  title: C2CMarket API
  version: 0.1.0
  summary: Backend contract foundation and official price lead slice.
  description: |
    This contract covers the first executable backend slice. C2CMarket is a price intelligence
    and community matching platform. It does not process payments, escrow funds, store third-party
    credentials, store plaintext native passwords, auto-deliver keys/accounts, or proxy upstream API
    traffic. Native C2CMarket account passwords are stored only as salted hashes.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
paths:
...
  /api/v1/reports:
    post:
      operationId: createReport
      tags: [Reports]
      security:
        - sessionCookie: []
          csrfToken: []
      summary: Submit a report for a contact snapshot, public user, carpool membership, or API purchase intent.
      description: User-submitted content must not contain full contact values, passwords, API keys, tokens, sessions, cookies, recovery codes, or other credential material.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Report submitted.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/me/appeals:
    post:
      operationId: createAppeal
      tags: [Reports]
      security:
        - sessionCookie: []
          csrfToken: []
      summary: Submit an appeal linked to a report or dispute.
      description: Appeal content must be non-sensitive and must not include credentials, tokens, full contact values, cookies, sessions, API keys, or recovery codes.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppealRequest"
      responses:
        "201":
          description: Appeal submitted.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      operationId: listMyAppeals
      tags: [Reports]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Current user's appeals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppealList"
        "401":
          $ref: "#/components/responses/Problem"
...
  /api/v1/admin/reports/{id}:
    get:
      operationId: getAdminReport
      tags: [AdminReports]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Admin report detail.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/admin/disputes/{id}:
    get:
      operationId: getAdminDispute
      tags: [AdminReports]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Admin dispute detail.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeCase"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/admin/appeals/{id}:
    get:
      operationId: getAdminAppeal
      tags: [AdminReports]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Admin appeal detail.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appeal"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: c2c_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Defaults to 20 and must be between 1 and 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor returned as nextCursor. Current implementation encodes an offset cursor and clients must not inspect it.
      schema:
        type: string
...
  responses:
    Problem:
      description: Problem Details error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RateLimited:
      description: Rate limit exceeded. Problem Details `code` is `RATE_LIMITED`.
      headers:
        Retry-After:
          schema:
            type: string
          description: Seconds until the current in-process rate-limit window resets when known.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    CreateReportRequest:
      type: object
      additionalProperties: false
      required: [targetType, targetId, reasonCode, title, description]
      properties:
        targetType:
          type: string
          enum: [contact_snapshot, public_user, carpool_membership, api_purchase_intent, api_order]
        targetId:
          type: string
          description: Target resource id, contact snapshot id, or reported username for public-user reports.
        targetLabel:
          type: string
          maxLength: 120
        reportedUsername:
          type: string
          description: Required for public-user reports and optional for contact snapshot reports.
        reasonCode:
          type: string
          enum: [invalid, unreachable, impersonation, other]
        title:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          minLength: 4
          maxLength: 1000
          description: Must not contain full contact values, passwords, API keys, tokens, sessions, cookies, recovery codes, or other credential material.
    Report:
      type: object
      required: [id, reporterUsername, reporterName, targetType, targetId, targetLabel, reportedUsername, reasonCode, title, status, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        reporterUserId:
          type: string
          format: uuid
          description: Admin response only.
        reporterUsername:
          type: string
        reporterName:
          type: string
        targetType:
          type: string
          enum: [contact_snapshot, public_user, carpool_membership, api_purchase_intent, api_order]
        targetId:
          type: string
        targetLabel:
          type: string
        reportedUsername:
          type: string
        reasonCode:
          type: string
          enum: [invalid, unreachable, impersonation, other]
        title:
          type: string
        description:
          type: string
          description: Admin response only; user-facing lists omit raw report description.
        status:
          type: string
          enum: [submitted, triaged, rejected, dispute_opened]
        adminReason:
          type: string
          description: Admin/self moderation context only; never part of public dispute summaries.
        handledByAdminId:
          type: string
          format: uuid
          description: Admin response only.
        handledAt:
          type: [string, "null"]
          format: date-time
        disputeId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    DisputeCase:
      type: object
      required: [id, targetType, targetId, targetLabel, primaryUsername, primaryDisplayName, counterpartyUsername, counterpartyName, status, publicSummary, publicResult, openedAt, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        targetType:
          type: string
          enum: [contact_snapshot, public_user, carpool_membership, api_purchase_intent, api_order]
        targetId:
          type: string
        targetLabel:
          type: string
        primaryUserId:
          type: string
          format: uuid
          description: Admin response only.
        primaryUsername:
          type: string
        primaryDisplayName:
          type: string
        counterpartyUserId:
          type: string
          format: uuid
          description: Admin response only.
        counterpartyUsername:
          type: string
        counterpartyName:
          type: string
        status:
          type: string
          enum: [open, waiting_info, resolved, closed]
        publicSummary:
          type: string
        publicResult:
          type: string
        adminReason:
          type: string
          description: Admin response only; never returned by public dispute endpoints.
        openedByAdminId:
          type: string
          format: uuid
          description: Admin response only.
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, "null"]
          format: date-time
        closedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    CreateAppealRequest:
      type: object
      additionalProperties: false
      required: [title, statement]
      properties:
        reportId:
          type: string
          format: uuid
        disputeId:
          type: string
          format: uuid
        targetType:
          type: string
          enum: [contact_snapshot, public_user, carpool_membership, api_purchase_intent, api_order]
        targetId:
          type: string
        title:
          type: string
          minLength: 2
          maxLength: 80
        statement:
          type: string
          minLength: 4
          maxLength: 1000
          description: Must not contain full contact values, passwords, API keys, tokens, sessions, cookies, recovery codes, or other credential material.
    Appeal:
      type: object
      required: [id, appellantUsername, appellantName, targetType, targetId, title, status, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        appellantUserId:
          type: string
          format: uuid
          description: Admin response only.
        appellantUsername:
          type: string
        appellantName:
          type: string
        reportId:
          type: string
          format: uuid
        disputeId:
          type: string
          format: uuid
        targetType:
          type: string
        targetId:
          type: string
        title:
          type: string
        statement:
          type: string
          description: Admin response only; public endpoints never expose appeal statements.
        status:
          type: string
          enum: [submitted, approved, rejected]
        adminReason:
          type: string
          description: Admin/self moderation context only.
        handledByAdminId:
          type: string
          format: uuid
          description: Admin response only.
        handledAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
    AppealList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Appeal"
        nextCursor:
          type: [string, "null"]
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string